+++ src/plone/patternslib/interfaces.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from zope.publisher.interfaces.browser import IDefaultBrowserLayer
-
-
-class IPlonePatternslibLayer(IDefaultBrowserLayer):
-    """Marker interface that defines a browser layer."""